+++ demisto_sdk/commands/content_graph/parsers/wizard.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-from typing import List
+from typing import List, Set
 
 from demisto_sdk.commands.common.constants import MarketplaceVersions
 from demisto_sdk.commands.content_graph.common import ContentType
@@ -17,8 +17,8 @@ class WizardParser(JSONContentItemParser, content_type=ContentType.WIZARD):
         self.playbooks: List[str] = self.get_playbooks()
 
     @property
-    def marketplaces(self) -> List[MarketplaceVersions]:
-        return [MarketplaceVersions.XSOAR]
+    def supported_marketplaces(self) -> Set[MarketplaceVersions]:
+        return {MarketplaceVersions.XSOAR}
 
     def get_packs(self) -> List[str]:
         packs: List[str] = []