+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Added support for marketplace-specific tags.
 * Fixed an issue where **download** would not run `isort`. @maxgubler
 * Fixed an issue where XSIAM Dashboards and Reports images failed the build.
+* Added support for **xpanse** marketplace to content graph.
 
 ## 1.7.7
 * Fixed an issue where paybooks **generate-docs** didn't parse complex input values when no accessor field is given correctly.